BUG#12762393 - CAN NOT SHOW CREATE VIEW IF UNDERLYING TABLES                ARE ALTERED
BACKGROUND:
There was refactoring done in 5.1, which changed the way the 
warnings are suppressed in the server. Before the refactoring,
the warnings were suppressed (or propagated to errors) based 
on the global flag.
Afterwards the warnings were suppressed/propagated by local 
instances of Internal_error_handler on a per case base.
This bug is an indication of an issue with that refactoring.
The issue was that ER_BAD_FIELD_ERROR was forgotten to be 
handled when opening the underlying tables in SHOW CREATE VIEW.

FIX:
As a fix for this bug, added a case to handle ER_BAD_FIELD_ERROR
in Show_create_error_handler to generate a warning instead of error 
when columns of view's underlying table is changed.
Also some redundant error codes have been removed in
Show_create_error_handler and in TABLE_LIST::hide_view_error()
by adding the appropriate comments.