An introductory change for the number 12794345 bug:
Let all platforms use the Perl variant of the script
"mysql_install_db".

This will later allow us to add setting a random password
for the root user (and other steps) to this script.

"scripts/mysql_install_db.sh" is now unused, cleanup (removal)
is left to a later change.