Bug #14731457  	MYSQL_INSTALL_DB FAILES WHEN DATADIR IS NESTED                 DIRECTORIES (WHICH DON'T EXIST)
The Perl variant of "mysql_install_db" does not create the path
leading to the datadir if it does not exist,
the shell variant did so (silently) by "mkdir -p".

The bug report contains the explanation why this implicit creation
is considered wrong, users should create such a non-standard path.
(Installation following the manual will never encounter this.)

If the script is called with a datadir option whose parent
directory does not exist, it will now give an error message.