WL#6402: Non blocking show slave status
SHOW SLAVE STATUS blocks when it's executed in parallel with STOP SLAVE,
what will block any monitoring tools that are checking replication
metadata. For this purpose is better to return not so coherent
information then no information at all.

To allow non blocking replication metadata checking a new option was
introduced "NONBLOCKING" to SHOW SLAVE STATUS, which when used returns
immediately instead of blocking when locks held on STOP SLAVE, this
command is intended for monitoring tools which requests should not
block.