Bug #14723291 - FTS AUXILIARY CREATE TABLE IGNORES FILE PER TABLE SETTING
The WL5980 changes changed the way srv_file_per_table is used, it was an
improvement over previous usage. However, this introduces a bug because the
FTS auxiliary tables are created using the internal SQL parser. We can't set
any special flags or influence the setting from SQL. We have to use the global
setting. The problem was that the AUX tables were being created in the System
Tablespace. Additionally orphaned tables could be left in the system tablespace
causing bloat and making it (close to) impossible to reclaim that space. This
fix addresses both problems. Create the AUX tables in separate tablespaces if
the srv_file_per_table flag is set and additionally checking for orphaned AUX
tables and removing them.

Drop FTS aux tables during DISCARD tablespace.

Several tests were fixed because they list files and the AUX files were
showing up in the directory listing after this fix.

rb://1370 Approved by Jimmy Yang.