Bug #14712962 - TRANSPORTING INNODB TABLESPACES WITH A FULL TEXT INDEX CRASHES MYSQLD
Server crashes when importing a table with FTS indexes. The fix is to mark
the FTS index as corrupt but allow the import of the data for the parent table.
The user can then drop and rebuild the FTS index. Also, flush the contents of
the parent table to disk even if the table has FTS indexes, when the user
issues:

	FLUSH TABLES t (WITH READ LOCK | FOR EXPORT);

However, we return a warning to the user that it is not fully supported.

With this change we partially support export/import of tables with FTS indexes.

Note: Importing a table that has FTS indexes without a .cfg file is undefined
behaviour and not recommended.

rb://1365 Approved by Jimmy Yang.