Bug #11761854: LEFT JOIN AND UNION PARSING ERROR
Replaced a bad parser hack with a slightly better one.

There is a test that checks that we don't put aliases after tables
within parentheses, e.g. SELECT 1 FROM (table) AS alias. This is not
legal syntax. At the same time, the test attempts to catch the case
when we have a non-query on the left-hand side of a nested union,
e.g. SELECT 1 FROM (table UNION SELECT 1).

Unfortunately, the test also fired for any union queries where the
right-hand side query term had a table within parentheses, e.g. SELECT
1 UNION SELECT 1 FROM (table) would raise an artificial parse error,
as it were, in error.

Fixed by splitting the test in two. The check for non-queries is moved
to the rule parsing nested union's. The test is put in a semantic
action that goes *last* in the rule. This is the proper place for
semantic actions in 99.9% of the cases. But since there was no
semantic action here before, bison obviously creates a somewhat
different PDA, hence some changed error messages in
subselect.test. The error message is still wrong, however - see Bug no
14731729 - so we care not.