Bug#14554568: CTRL-C BEHAVIOR VIOLATES PRINCIPLE OF               LEAST ASTONISHMENT
MySQL client's current behaviour was to kill the last
query on receipt of SIGINT (Ctrl + C). In case no query
was found to be executing, the process itself would
be terminated.

Going with the POLA, now, 'Ctrl + C' will send a 'KILL
QUERY' to the server if a query is found to be executing.
This, however can be avoided by using --sigint-ignore
option.

'Ctrl + C' would now also cause the current line to
terminate similar to \c command (invariable to the use
of --sigint-ignore option). Lines, thus cancelled, would
not make into client's history.

'Ctrl + D' on *NIX systems would work as it did (EOF), to
terminate the client when entered as the very first character
on the prompt. However, with this patch the client would show
same behaviour on Windows when 'Ctrl + Z' is used.