Bug#14733674 CREATE UNIQUE INDEX CREATES NON-UNIQUE INDEX (take 3)
row_log_apply_op_low(): When updating a record, check both low_match
and up_match for duplicate, no matter if one of the records is
delete-marked. When applying ROW_OP_INSERT, check both up_match and
low_match, and drop the condition on the cursor position. This is a
follow-up to rb:1385.

With this change, we may be introducing false duplicate key errors.
More work needs to be done in this area before GA release.

rb:1387 approved by Jimmy Yang