Fix for bug #14723364 "PRINT_KEYDUP_ERROR PRINTS DEFAULT VALUE INSTEAD OF NULL".
In-place implementation of ALTER TABLE for InnoDB storage engine
in some special cases can emit duplicate key error for NULL key
values. When such a thing has happened NULL value was not properly
shown in duplicate key error message, instead default value for
column's type was shown.

The problem was that KEY_PART_INFO structure, which describes key
and which is used for printing key values by print_keydup_error()
function, didn't have KEY_PART_INFO::null_offset/null_bit properly
set. As result all such fields were treated as non-NULL and even
though InnoDB layer properly set column value in a record to NULL,
print_keydup_error() interpreted as a non-NULL value.
Note that KEY_PART_INFO structure which is used in this case is
specially constructed for in-place ALTER TABLE and describes parts
of key in new version of table. KEY_PART_INFO which is normally used
for other statements and which is constructed when table is opened
is not affected by this code.

This fix solves the problem by changing code in InnoDB "handler"
layer to properly set KEY_PART_INFO::null_offset/null_bit members,
along with KEY_PART_INFO::field member.

There is no separate test case in this patch as to reproduce the
problem one has to either rely on using DBUG_EXECUTE_IF macro or
on buggy behavior which is likely to be changed in near future.