Worklog #6073: Remove INSERT DELAYED: Fourth patch, remove delayed locks.
This patch removes the lock type TL_WRITE_DELAYED from the list of
lock types. The locking algorithms are trivially modified. Equal
comparisons with TL_WRITE_DELAYED are removed as either false or true,
while greater/less than comparisons are shifted to compare with the
next higher or lower value. If() statements based on such comparisons
are modified so that either if() is removed - if it's always true - or
the whole statement including body if always false. References in
documentation to TL_WRITE_DELAYED are also deleted.

A few functions are rendered obsolete after this change:

- thr_reschedule_write_lock()
- upgrade_lock_type()
- create_insert_stmt_from_insert_delayed()