Bug #14726722: "MYSQL_SECURE_INSTALLATION" IS INCOMPATIBLE WITH THE "PASSWORD EXPIRED" FLAG
mysql_secure_installation tries to be smart and update all the root account's
passwords through manipulating the system tables directly and then 
re-reading them.
But since these statements are not in the list permitted in password expired
mode this causes errors if the root account used to connect is marked as
expired.
Fixed by first issuing a SET PASSWORD statement for the connecting account
to lift off the expiration flag for it.
Also extended the mysql.user UPDATE statements to reset the expiration flag
for all other root accounts.