Bug#14601664 ASSERTION FAILED: *REF && (*REF)->FIXED, FILE ITEM.CC
Problem: A query with a subquery with an outer reference and a
subquery with the same outer reference causes an assertion to fire.

The outermost query has an Item_field that is referred to by the inner
queries. The outer reference in the middle query resolves to this
field, and the Item_field in the middle query is replaced by an
Item_outer_ref. This Item_outer_ref is added to the inner_refs_list by
Item_field::fix_outer_field() so that it can be fixed later by
fix_inner_refs(). The outer reference in the innermost query resolves
to the middle query's Item_outer_ref, but an assertion fires in
Item_field::fix_outer_field() or Item_ref::fix_fields(), depending on
the type of the item in the innermost query, since the referenced item
is not fixed yet.

Fix: Open up assertions to allow references to unfixed
Item_outer_refs, similar to what was done in
resolve_ref_in_select_and_group() by patch for bug #13735050.