Bug #14698309: 'MYSQLADMIN PASSWORD' IS INCOMPATIBLE WITH                THE "PASSWORD EXPIRED" FLAG
mysqladmin ... password ... was expecting that it can always execute
commands other than SET PASSWORD against the server. And it was 
using this opportunity to try to be smart and detect if "old" pre-4.1 
passwords are being used by this server.
If any of these commands failed it was just giving up.
This is obviously incompatible with password expiration, as the only
accepted command in this state would be SET PASSWORD. And this
renders mysqldump .. password .. unusable as an option to update
the password for such accounts.
Fixed mysqldump to detect this particular error code sent by the 
password expiration check and react by skipping the 
auto-adjustment (and logging an warning about it if run with 
--verbose) and trying its best according to the parameters supplied
to it.

Added a testcase.