Bug#14756089: ALTER TABLE: ASSERTION FAILED: RECLENGTH ==               FIELD->OFFSET + DATA_OFFSET
This assertion was triggered if ALTER TABLE was used to rename
a column to same name as an existing column while also reordering
the renamed column using AFTER/FIRST.

Constructing the new list of columns during ALTER TABLE is a two
step process. First the new column are made and then they are
reordered (if necessary). During the reordering, columns that should
change positions are first removed from the list and then added back
at the proper position.

The bug was that when removing columns we matched on column name.
So if the column to be reordered had the same name as an existing
(unchanged) column, we could end up removing the wrong column.
This would lead to problems (including triggering the assert) later on.
Note that all this happens before we check for duplicate column names.

This patch fixes the problem by instead matching the columns using
pointer equality since the two lists using for matching will have
the same element in case of column change (such as reorder).

This bug was a regression from the patch for Bug#11748057.