Bug#14644842: ASSERT 'INITED == NONE || (INITED == RND && SCAN)'               IN HANDLER::HA_RND_INIT
The problem was for a query with dynamic range access. For
each row combination, it was evaluated whether to use
range access or index scan. Each time a range scan was chosen,
the chosen index was inited, read and finally closed the 
next time access type was chosen. However, when index scan was
chosen instead, the index was not closed before next iteration.
Thus, if a range scan was chosen after an index scan, a 
DBUG ASSERT triggered during initialization of the range scan.

The fix is to close the index before each new iteration.