BUG#14683107 - TRUNCATED VALUE FOR THREAD_CACHE_SIZE DIFFERS                WHILE SERVER STARTUP
DESCRIPTION:
When mysqld is started with a negative thread_cache_size 
parameter, the thread_cache_size Global variable is reset
to 16384 (MAX SIZE).

Whereas when thread_cache_size is manually set using
SET GLOBAL command with a negative value, the value of 
"thread_cache_size" is reset to 0, which is an inconsistent
behaviour.


ANALYSIS :
In case of manually setting "thread_cache_size" using
SET GLOBAL command, sys_var::do_check() takes care of 
sanitizing a negative input to 0 on unsigned variables, hence
"thread_cache_size" is reset to 0.

However during startup "thread_cache_size" is initialized by 
getopt_ull() passing the input string prefixed by '-' sign.
This is then passed to eval_num_suffix() which uses strtoll()
to convert string to numeric value. This disregards the return
type and return 0xFFFFFFFFFFFFFFFF for -1 and hence this is
assigned to the Global variable and it is truncated to 16384.


FIX :
getopt_ull() now checks to see if the input argument contains
a '-' sign and if so 0 is returned.