Fix for Bug#14682735 UNNECESSARY "USING WHERE" IN EXPLAIN RESULT
For queries using ref access on string data types, the ref access
condition was evaluated also as part of the query condition or pushed
down as an index condition to the storage engine. For most data types
this extra evaluation of the ref access condition is unnecessary and
adds overhead.

For multiple data types (eg. integers, double) the extra evaluation of
the ref condition was already avoided. This patch extends this to
avoid the extra evaluation of the ref access condition for all data
types with the exception of float (see comments in the code about
float).