Bug#14742760 SELECT WITH 12MIO COLUMNS CRASHES MYSQL                    CLIENT
MySQL client internally used alloca() to allocate
some memory from stack to keep a track of numeric
fields. This allocation is unsafe and might result
in stack overrun if the number of fields is very
big.

Fixed by adding a logic to switch to my_malloc instead,
if the requested size exceeds certain limit.

Tested the patch manually for 12M fields.