Bug#14788976 ASSERTION FAILED: (CREATE_INFO->OPTIONS & 1) ||              THD->MDL_CONTEXT.IS_LOCK_OWNER(M
The triggered assertion checks that we have an exclusive metadata
lock on the table created by CREATE TABLE ... LIKE in cases where
the statement completes without error and the table created is not
a temporary table.

It was triggered if CREATE TABLE IF NOT EXISTS ... LIKE was used
in LOCK TABLES mode. Because of IF NOT EXISTS, the statement
will not fail even if the table did already exist. Also because
of LOCK TABLES mode we will not have an exclusive metadata lock
on the table at this point.

This patch fixes the problem by adjusting the assert to take
the LOCK TABLE possibility into account. This bug was only 
present in debug builds, release builds had no issues.