Bug#14117018 - MYSQL SERVER CREATES INVALID VIEW DEFINITION
Analysis:
---------
If CREATE VIEW has unnamed column in SELECT query, 

For Example:
  SELECT '' AS DUAL;

then while creating the view, name will be generated for
such column.

If the CREATE VIEW has UNION in SELECT queries:

  SELECT_1 UNION SELECT_2;

then while creating the view if there is a need to generate 
column_name then it was generated for only SELECT_1. And such 
columns in SELECT_2 are not processed. Because of this, view is 
created with the invalid query. Re-creating view using output 
of mysqldump or "SHOW CREATE VIEW" command will fail.

Fix:
---------
As a fix, modified code to consider all SELECTs of UNION to 
check if auto generated column names are conforming and to 
generate conforming name if necessary.