Bug#14776799: INNODB-WL5980-DISCARD AND I_MAIN.ALTER_TABLE TIMEOUT ON PB2
The timeout occurred on Windows because a temporary tablespace file
was not being deleted from the fil_space dictionary.

The testcase that was failing inside i_main.alter_table.test is moved
out to its own test using the original bug number since tests with
IMPORT and DISCARD should not be run in embedded mode.

In addition, this patch fixes a failure within i_main.alter_table.test
which has been failing with this error;

mysqltest: At line 147: command "copy_file" failed with error 1. my_errno=17

That error does not occur anymore in the new test file because it has
a --remove-file to delete the backed up .cfg and .ibd files.

Code change provided by Sunny and reviewed approved by marko in
http://bur03.no.oracle.com/rb/r/1436/.