Bug #14520559 - INNODB; LARGE BLOB INSERTS ARE SLOW AND TROUBLESOME
Large BLOBS are slow with debug builds due to the large number of
calls to mtr_memo_contains() and the extra large size of the main
MTR.  Non-debug builds do not seem to suffer noticeably. That
problem is not addressed by this patch.

This patch fixes the 'troublesome' assert that happens with 4k page
size and with some Compressed tablespaces.  The assert is described
fully in the bug report.  It has to do with how a system tablespace
is initially expanded.  At 32Mb, it is normally expanded by more than
one extent at a time.  But with 4k pages, that must happen sooner.

http://bur03.no.oracle.com/rb/r/1301/ approved by Marko.