Patch for Bug#14788443: ASSERTION FAILED: M_CPP_BUF <= PTR && PTR <= M_CPP_BUF + M_BUF_LENGTH.
The bug was a regression in 5.6.3 introduced by a patch for 12546960
(revid: magne.mahre@oracle.com-20110725105406-kwpzpksmrsjgeloe).

The problem was that in case of syntax error, the lexer was left in
non-consistent state (Lex_input_stream::m_cpp_ptr was poiting
to after-ending-zero byte).

The fix is to return (unget) one character to the Lex_input_stream in
case of syntax error.