Follow-up to Bug#14712710 BOGUS DUPLICATE KEY ERROR DURING ONLINE ALTER TABLE
row_merge_read_clustered_index(): Ignore the old version of a record
if it is delete-marked. This can happen when a DELETE is committed,
but purge does not remove the record before an INSERT of the same
PRIMARY KEY value arrives. In this case, the INSERT would be performed
by updating the delete-marked record.

The original fix was rb:1438 (revision-id
marko.makela@oracle.com-20121024123439-rapozfh9ghvfdtk8).

Approved by Inaam Rana (on IM).