Bug#11753308: MYSQL_KILL() AND COM_PROCESS_KILL ONLY HANDLE 32-BIT IDS
The connection ID / "thread-ID" is currently 32 bit on the wire, but may be 64 bit in the API,
depending on the definition of ulong on a given platform. If we silently drop half the
bits on the floor, mysql_thread_id() in Connector-C may give incorrect results, and mysql_kill()
might ask for the wrong (and usually long expired) thread to be terminated, even if a 64-bitly
correct ID were obtained with "SELECT connection_ID()".

We now guard against some 32/64 issues.
We do this in both the client and the server to address mixed scenarios of old server /
new client and vice versa.

Using SQL instead of the C API (SELECT CONNECTION_ID() / KILL CONNECTION ...) should
produce better results. mysqladmin already does this.

Docs implications:

- Discourage use mysql_thread_id() / mysql_kill(). current status is that they produce incorrect
  results / try to kill the wrong thread once more than 2^32 threads are created (IOW, this only
  affects extremely busy servers or those with an extreme uptime. The server also needs to use
  64-bit thread_IDs for this to apply. This is platform specific, not a mysqld configurable.)
  Patched status is that the calls will work normally as long as thread_IDs are in safe range
  (< 2^32), but will throw an error a) if mysql_kill() in Connector-C is called with a value
  >= 2^32, or b) if mysql_kill() is used at all, even with ostentably correct values, once
  thread_ID in the server has exceeded what can be expressed within 32 bits.

- encourage use of the SQL equivalents, SELECT connection_ID() / KILL CONNECTION ...

- mysqladmin already does this, and is therefore safe.

- mysql_kill() will throw CR_INVALID_CONN_HANDLE if it is given a larger value than can be
  expressed in the 32-bit wire protocol.

- mysqld will throw ER_DATA_OUT_OF_RANGE on mysql_kill() once thread_ID in mysqld exceeds
  32-bits.

Backport from 5.7.