Bug#14505672 MYSQL_CONFIG_EDITOR REMOVE --ALL FAILS Bug#14545999 MYSQL_CONFIG_EDITOR SET -ALL BAD BEHAVIOR Bug#14545989 MCE DOES NOT WARN ON WRITE FAILURE
This patch modifies the way options are handled by MCE.
The options now are divided and handled in 2 different
categories. The 1st set of options are program options
while the 2nd set being specific to MCE commands. In
order to bring in support for commands in the option
handling routine, handle_options(), some extra logic
has been added to stop option processing when a non-
options string matches with a command specified through
the new command_list parameter.

Also, added the support for --user, --password, --host
options for remove command, which could not be added
before due to the option handling design constraints.

Added test cases to test the added functionalities.