Bug#11754168 - PARTS OF INDEX_MERGE_INNODB.TEST ARE DISABLED                DUE TO EXPLAIN DIFFS
When a huge amount of rows are updated, InnoDB's #records
estimates go crazy for a while because indexes are full of 
delete-marked rows. InnoDB does not understand that 
delete-marked rows are deleted so they are counted. To 
improve testing, the CREATE TABLE - INSERT sequence is 
modified to not update rows. The result is that a bunch of 
tests now use index_merge where table scan used to be 
performed. 

There are no differences to the tests other than the way the
tables are created and populated.