Bug#14789301: CRASHING SERVER BY STORED FUNCTION REFERENCING USER DEFINED VARIABLE IN QUERY
There are 2 steps resulting in the crash:
1) the stored procedure was executed during JOIN::prepare resulting
    in setting thd->user_var_events_alloc to NULL (due to LTM_NONE)
2) the next time the stored procedure was executed in the same statement
    it crashes due to thd->user_var_events_alloc was null. (it was not
   set to thd->mem_root since LTM_PRELOCKED).

The fix is to:
* avoid executing the stored program during prepare phase of
Item_func_isnull (update_used_tables).

Also updated a comment which got out of date after bug 14247298.
Also added test case copied from the duplicate bug:
BUG#11765560 - SEGFAULT ON SHOW TABLE STATUS (MYSQLDUMP) OF NESTED
VIEWS
which is a duplicate in 5.6 (other cause in 5.1/5.5), but with another crash.
In 5.1/5.5 this patch cannot be used since it lacks of with_stored_program,
and the problem is caching strategy for is_not_null/isnull and
DETERMINISTIC stored programs.