BUG#14797808: GTIDS AND BACKUP DON'T WORK TOGETHER
Background
----------
The GTID feature was introduced in 5.6. It currently does not work with
our backup tools MEB, mysqldump, and mysqlhotcopy: these tools are not
capable to restore the GTID state.

Suppose we perform the following steps:
1. B := backup from server_1
2. restore B on server_2
3. replicate from server_1 to server_2

Problem
-------
The desired behavior is that replication starts from the point where
the backup was taken. However, currently it starts from the beginning
of history.

Solution
--------
On server core, GTID_PURGED must be settable. It is only possible to
*add* GTIDs to GTID_PURGED and only when server is a new server
(GTID_EXECUTED == '').