Bug#14838882: INIT_IO_CACHE() HAS SUSPICIOUS MEMSET()
The problem was that init_io_cache() contained a memset() that with
SAFE_MUTEX was used to clear a mutex - IO_CACHE::append_buffer_lock.
However the size of memory cleared was the size of the IO_CACHE pointer
rather than the size of the IO_CACHE::append_buffer_lock mutex.

This patch fixes the problem by clearing the appropriate number
of bytes.