BUG#11763004 - _DOSMAPERR WIPES OUT THE ERRNO SET IN                _OPEN_OSFHANDLE AND GIVES (ERRNO: 2
DESCRIPTION:
When an API or function call fails in mysqld.exe under windows,
Error Logs report Error Code 22 for certain cases.

ANALYSIS :
When an error occurs my_osmaperr() is called with the value 
from GetLastError(). The mapping which exists in my_winerr.c 
is fairly limited and EINVAL (22) is returned for a lot 
Windows Error Codes, which makes it hard to detect the cause 
from error Logs.

Reproducing an issue which returns an unmapped Windows Error
Codes is very difficult, hence an test case for this issue hs
intentionally been left out.

FIX :
my_osmaperr() now saves the original GetLastError() value in
thr_winerr. This saved value is used again in my_strerror()
incase the nr value is EINVAL (22) to give a better error
message.