Script "mysql_install_db" could run into trouble signalled by       tr: Input file error: Illegal byte sequence (taken from a Solaris test) which then led to       Could not generate a random password - not setting one
The reason is that "tr" (used to ensure the password is
alphanumeric, important to avoid issues in pasting)
may need LC_CTYPE to be set, will not understand the
"alnum" character class otherwise.

The fix is to specify "LC_CTYPE=C" when calling "tr".
It is done for this call only, will not affect user settings
for any other program.

Detected during internal testing on Solaris, bug #14848215.