Bug #14680057 	NDB : LCP FRAG WATCHDOG GIVES ERROR WITH V HIGH TABLE, FRAGMENT AND ROW COUNTS
The recently added LCP fragment scan watchdog occasionally reported
problems with LCP fragment scans with very high table id, fragment id,
and row count values.

This was due to the watchdog not accounting for the time spent draining
the 'Backup' buffer used to buffer rows before writing to the fragment
checkpoint file.

This is fixed so that in the final stage of an LCP fragment scan, the 
watchdog switches from monitoring rows scanned, to monitoring the 
buffer size in bytes.  The buffer size should decrease as data is 
written to the file, after which the file should be promptly closed.