Fixed merge problems between mysql-5.6 and mysql-trunk.
The error ER_READ_ONLY_MODE was not numbered the same way in 5.6 and trunk (5.7),
causing discrepancies in error numbering.

This is a bug, as it impacts compatibility between 5.6 and later versions.

This fix moved ER_READ_ONLY_MODE from the "new in 5.7" to the "new in 5.6" section,
because this error is also present in 5.6.

As a mechanical result, the following errors have been renumbered:

#define ER_READ_ONLY_MODE 1836
#define ER_GTID_NEXT_TYPE_UNDEFINED_GROUP 1837
#define ER_VARIABLE_NOT_SETTABLE_IN_SP 1838
#define ER_CANT_SET_GTID_PURGED_WHEN_GTID_MODE_IS_OFF 1839
#define ER_CANT_SET_GTID_PURGED_WHEN_GTID_EXECUTED_IS_NOT_EMPTY 1840
#define ER_CANT_SET_GTID_PURGED_WHEN_OWNED_GTIDS_IS_NOT_EMPTY 1841
#define ER_GTID_PURGED_WAS_CHANGED 1842
#define ER_GTID_EXECUTED_WAS_CHANGED 1843
#define ER_BINLOG_STMT_MODE_AND_NO_REPL_TABLES 1844
#define ER_FILE_CORRUPT 1845
#define ER_ERROR_ON_MASTER 1846
#define ER_INCONSISTENT_ERROR 1847
#define ER_STORAGE_ENGINE_NOT_LOADED 1848
#define ER_GET_STACKED_DA_WITHOUT_ACTIVE_HANDLER 1849
#define ER_WARN_LEGACY_SYNTAX_CONVERTED 1850
#define ER_ERROR_LAST 1850

Test scripts have been adjusted accordingly.