BUG#13520710: INCORRECT VIEW CREATION WITH WHERE CLAUSE
Analysis:
--------
The 'WHERE' clause contains incorrect values for string literals
in the query section of the view definition when the charset
defined by 'character_set_client' and 'character_set_connection'
are different.

While parsing the query, the string literals are converted from the
charset defined by 'character_set_client' to the charset defined by
'character_set_connection'.

During the formation of the query section of the view definition,
the 'string literals' were printed in the same charset as that of
the 'charset_set_connection' without converting it to
'character_set_client'. This resulted in incorrect string values.
[The query section of the view definition is stored in the same
charset as defined by character_set_client].

Fix:
---
a) A new query type 'QT_TO_ARGUMENT_CHARSET' is added to determine
whether to convert the string literal to 'character_set_client'.
b) A new method 'Item_string::convert_and_print' handles the conversion
of the string from 'character_set_connection' to 'character_set_client'
and then prints the string literals which forms the query section of
the view definition. Thus ensuring correct string literal values in
the 'WHERE' clause.

Other issue fixed with this patch:
----------------------------------
SHOW CREATE PROCEDURE proc_name contains incorrect values for string
literals in the SET statement. This occurs when 'character_set_client' 
and 'character_set_connection' are different when the procedure is
defined.